post detail component is created
